WebApi| updated CustomerMovieOperations - MovieOperations - Entities - DataGenerator

+++ WebApi/Application/MovieOperations/Commands/UpdateMovie/UpdateMovieCommand.cs
@@ -41,7 +41,7 @@ public class UpdateMovieModel
 {
     public string? Name { get; set; }
     public int? Year { get; set; }
-    public decimal? Price { get; set; }
+    public double? Price { get; set; }
 
     public int? GenreId { get; set; }    
     public int? DirectorId { get; set; }
